requirements.txt
asgiref==3.7.2
Django==4.2.1
django-smart-selects==1.6.0
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.3
